Implemented functionality and Design UI for Rating & Review Section on Product Screen

+++ lib/models/product_ratings.dart
@@ -13,7 +13,7 @@ class ProductRatings {
   ProductRatings.empty();
 
   String? productId;
-  int? avgRating;
+  double? avgRating;
   int? totalNumberOfRating;
   List<Rating>? ratings;
